Delete grade

+++ ADASOFT/ADASOFT/Data/Entities/Grade.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public StudentCourse StudentCourse { get; set; }
 
-        public float Percentage { get; set; }
+        public decimal Percentage { get; set; }
 
-        public float Grades { get; set; }
+        public decimal Grades { get; set; }
         //public ICollection<FinalGrade> Grades { get; set; }
