+++ ADASOFT/ADASOFT/Models/GradeViewModel.cs
@@ -7,9 +7,9 @@ namespace ADASOFT.Models
         public int Id { get; set; }
         public int StudentCourseId { get; set; }
 
-        public float Percentage { get; set; }
+        public decimal Percentage { get; set; }
 
-        public float Grades { get; set; }
+        public decimal Grades { get; set; }
         //public ICollection<FinalGradeViewModel> Grades { get; set; }
 
     }
